use internal merge tool
http://stackoverflow.com/questions/9003588/how-resolve-this-mercurial-conflict

If you don't like to see merge tools pop up, then you can set

[ui]
merge = internal:merge

to make Mercurial use the internal three-way merger only. It will merge files fine when the edits are not in conflicts. When there is a conflict, the file is marked as "unresolved" and conflict markers are stored in the file.

You then need to edit the file by hand to get the version you want. You can also re-merge and pick either the local (your) version:

$ hg resolve --tool internal:local your-file
or the other version:

$ hg resolve --tool internal:other your-file